+++ src/helpers/ScryptEncryption.test.ts
